+++ deploy/10-deploy-rsETH-oracles.ts
@@ -7,20 +7,9 @@ import { ADDRESSES } from "../helpers/deploymentConfig";
 const func: DeployFunction = async function ({ getNamedAccounts, deployments, network }: HardhatRuntimeEnvironment) {
   const { deploy } = deployments;
   const { deployer } = await getNamedAccounts();
-  const proxyOwnerAddress = network.live ? ADDRESSES[network.name].timelock : deployer;
-  let WETH;
-  let rsETH;
-  if (network.name === "sepolia" || network.name === "ethereum") {
-    ({ WETH } = ADDRESSES[network.name]);
-    rsETH =
-      network.name === "sepolia"
-        ? "0xfA0614E5C803E15070d31f7C38d2d430EBe68E47"
-        : "0xA1290d69c65A6Fe4DF752f95823fae25cB99e5A7";
-  }
-  const defaultProxyAdmin = await hre.artifacts.readArtifact(
-    "hardhat-deploy/solc_0.8/openzeppelin/proxy/transparent/ProxyAdmin.sol:ProxyAdmin",
-  );
-  console.log(`Timelock (${proxyOwnerAddress})`);
+  const proxyOwnerAddress = ADDRESSES[network.name].timelock;
+  const { WETH, rsETH } = ADDRESSES[network.name];
+
   const redStoneOracle = await hre.ethers.getContract("RedStoneOracle");
   const resilientOracle = await hre.ethers.getContract("ResilientOracle");
   const chainlinkOracle = await hre.ethers.getContract("ChainlinkOracle");
@@ -34,10 +23,6 @@ const func: DeployFunction = async function ({ getNamedAccounts, deployments, ne
     proxy: {
       owner: proxyOwnerAddress,
       proxyContract: "OptimizedTransparentProxy",
-      viaAdminContract: {
-        name: "DefaultProxyAdmin",
-        artifact: defaultProxyAdmin,
-      },
     },
     skipIfAlreadyDeployed: true,
   });
@@ -51,16 +36,11 @@ const func: DeployFunction = async function ({ getNamedAccounts, deployments, ne
     proxy: {
       owner: proxyOwnerAddress,
       proxyContract: "OptimizedTransparentProxy",
-      viaAdminContract: {
-        name: "DefaultProxyAdmin",
-        artifact: defaultProxyAdmin,
-      },
     },
     skipIfAlreadyDeployed: true,
   });
 };
 
 func.skip = async () => hre.network.name !== "ethereum" && hre.network.name !== "sepolia";
-
 func.tags = ["rsETHOneJumpOracles"];
 export default func;