+++ hardhat.config.ts
@@ -87,6 +87,20 @@ const config: HardhatUserConfig = {
           },
         },
       },
+      {
+        version: "0.5.16",
+        settings: {
+          optimizer: {
+            enabled: true,
+            runs: 200,
+          },
+          outputSelection: {
+            "*": {
+              "*": ["storageLayout"],
+            },
+          },
+        },
+      },
     ],
   },
   networks: {